added update order capability

+++ src/main/java/de/arkem/clean/arc/refactoring/demo/legacy/garage/portal/database/OrderDataCrudRepository.java
@@ -2,6 +2,6 @@ package de.arkem.clean.arc.refactoring.demo.legacy.garage.portal.database;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderDataCrudRepository extends CrudRepository<OrderDataDbo, Integer> {
+public interface OrderDataCrudRepository extends CrudRepository<OrderDataDbo, Long> {
 
 }
